+++ lang/index.js
@@ -1,0 +1,10 @@
+module.exports = {
+  OPEN_ADDR: '打开 url', 
+  CONFIG_SAVED: '配置已保存',
+
+  CONFIG_NOT_EXISTS: 'yyl.config 路径不存在',
+  CONFIG_PARSE_ERROR: '配置解析错误',
+  MISS_NAME_OPTIONS: '缺少 --name 属性',
+  NAME_OPTIONS_NOT_EXISTS: '--name 属性设置错误',
+  CONFIG_ATTR_IS_NEEDFUL: 'config 中以下属性为必填项'
+}