+++ lib/yh.d.ts
@@ -1,27 +1,30 @@
-import { userInfo } from "os";
-import { isInterfaceDeclaration } from "@babel/types";
+import { YylConfig, Env } from 'yyl-config-types'
 
 export default class Handler {
-  constructor({ log: logFunction, vars: IVars });
+  constructor({ log: LogFunction, vars: Vars });
   /** vars */ 
-  vars: IVars;
+  vars: Vars;
   /** 设置 vars */
-  public setVars(vars: IVars): IVars;
+  public setVars(vars: Vars): Vars;
   /** 去掉协议 */
   public hideProtocol(iPath: string): string;
   /** 语法糖 replace */
   public sugarReplace(str: string, alias: object): string;
   /** config 格式化 */
-  public parseConfig(configPath: string, iEnv: IEnv, returnKeys: string[] | string): Promise<any>;
+  public parseConfig(configPath: string, env: Env, returnKeys: string[] | string): Promise<any>;
   /**  构建相关操作 */
   public optimize: Optimize
 }
 
-type logFunction = (type: string, status: string, args: any[]) => void;
+/** log 输出 function 配置 */
+export type LogFunction = (type: string, status: string, args: any[]) => void;
 
-interface Optimize {
+/** script 类型 */
+export type ScriptType = 'watch' | 'all'
+
+export interface Optimize {
   /** 初始化 */
-  init(config: Config, iEnv: IEnv): void;
+  init(config: YylConfig, env: Env): void;
   /** 初始化 插件操作 */
   initPlugins(): Promise<any>;
   /** 打开 home page */
@@ -30,18 +33,13 @@ interface Optimize {
   livereload(): Promise<any>;
   /**  报错配置到 服务器操作 */
   saveConfigToServer(): Promise<any>;
+  /** 初始化 config.[all|watch].beforeScripts */
+  initBeforeScripts(ctx: ScriptType): Promise<any>;
+  /** 初始化 config.[all|watch].afterScripts */
+  initAfterScripts(ctx: ScriptType): Promise<any>;
 }
 
-interface IEnv {
-  name?: string;
-  remote?: boolean;
-  proxy?: boolean;
-  isCommit?: boolean;
-  workflow?: string;
-  ver?: 'remote';
-}
-
-interface IVars {
+export interface Vars {
   /** 本程序根目录 */
   BASE_PATH?: string;
   /** init path */
@@ -75,58 +73,3 @@ interface IVars {
   /** yyl 版本 */
   PKG_VERSION: string;
 }
-
-// + 配置相关
-interface Config {
-  /** seed 包名称 */
-  workflow: string;
-  /** seed sub name */
-  seed?: string;
-  /** yyl 版本 */
-  version: string;
-  /** 平台 */
-  platform: configPlatform;
-  proxy?: IProxy;
-  localserver: ConfigLocalserver;
-  dest: ConfigDest;
-  commit: ConfigCommit;
-  concat?: object;
-  resource?: object;
-  plugins?: string[];
-  webpackConfigPath: string;
-  alias: object;
-}
-
-interface IProxy {
-  port: number,
-  localRemote?: object;
-  homePage?: string;
-}
-
-interface ConfigLocalserver {
-  root: string;
-  port: number;
-}
-
-interface ConfigDest {
-  basePath: string;
-  jsPath: string;
-  jslibPath: string;
-  cssPath: string;
-  htmlPath: string;
-  imagesPath: string;
-  tplPath: string;
-  revPath: string;
-}
-
-interface ConfigCommit {
-  type: configCommitType;
-  revAddr: string;
-  hostname: string;
-  staticHost?: string;
-  mainHost?: string;
-}
-
-declare type configPlatform = 'pc' | 'mobile';
-declare type configCommitType = 'svn' | 'gitlab-ci';
-// - 配置相关